# [Rust] Use flattened list