Main: `| 23.60 ms  | 341.14 ms  | 3.05 s `

This PR:
```
Rust:

        Benchmark 1: ./target/release/rust
        Processing time (w/o IO): 21.976574ms
        Processing time (w/o IO): 22.039975ms
        Processing time (w/o IO): 21.961573ms
        Processing time (w/o IO): 21.942074ms
        Processing time (w/o IO): 21.971674ms
        Processing time (w/o IO): 21.965574ms
        Processing time (w/o IO): 22.814792ms
        Processing time (w/o IO): 22.23478ms
        Processing time (w/o IO): 23.252801ms
        Processing time (w/o IO): 21.891372ms
        Processing time (w/o IO): 22.029076ms
        Processing time (w/o IO): 22.067376ms
        Processing time (w/o IO): 21.951973ms
          Time (mean ± σ):      36.8 ms ±   1.6 ms    [User: 31.4 ms, System: 5.3 ms]
          Range (min … max):    35.4 ms …  39.9 ms    10 runs
         
Rust:

        Benchmark 1: ./target/release/rust
        Processing time (w/o IO): 319.851499ms
        Processing time (w/o IO): 320.719118ms
        Processing time (w/o IO): 321.126927ms
          Time (mean ± σ):     379.1 ms ±   0.4 ms    [User: 353.6 ms, System: 25.5 ms]
          Range (min … max):   378.8 ms … 379.4 ms    2 runs
         
Rust:

        Benchmark 1: ./target/release/rust
        Processing time (w/o IO): 2.816425475s
        Processing time (w/o IO): 2.810520034s
        Processing time (w/o IO): 2.845213766s
          Time (mean ± σ):      3.021 s ±  0.027 s    [User: 2.949 s, System: 0.070 s]
          Range (min … max):    3.002 s …  3.040 s    2 runs

```